fix: move solidity test for OP to op.ts
