+++ crates/edr_napi/test/solidity-tests.ts
@@ -1,21 +1,7 @@
 import { assert } from "chai";
 
-import {
-  Artifact,
-  ArtifactId,
-  ContractData,
-  EdrContext,
-  L1_CHAIN_TYPE,
-  SolidityTestRunnerConfigArgs,
-  SuiteResult,
-  l1SolidityTestRunnerFactory,
-  // HACK: There is no way to exclude tsc type checking for a file from the
-  // CLI, so we ignore the error here to allow `pnpm testNoBuild` to pass.
-  // @ts-ignore
-  OP_CHAIN_TYPE,
-  // @ts-ignore
-  opSolidityTestRunnerFactory,
-} from "..";
+import { EdrContext, L1_CHAIN_TYPE, l1SolidityTestRunnerFactory } from "..";
+import { loadContract, runAllSolidityTests } from "./helpers";
 
 describe("Solidity Tests", () => {
   const context = new EdrContext();
@@ -24,11 +10,6 @@ describe("Solidity Tests", () => {
     await context.registerSolidityTestRunnerFactory(
       L1_CHAIN_TYPE,
       l1SolidityTestRunnerFactory()
-    );
-
-    await context.registerSolidityTestRunnerFactory(
-      OP_CHAIN_TYPE,
-      opSolidityTestRunnerFactory()
     );
   });
 
@@ -144,88 +125,4 @@ describe("Solidity Tests", () => {
       }
     }
   });
-
-  it("executes tests for OP chain", async function () {
-    const artifacts = [
-      loadContract("./artifacts/SetupConsistencyCheck.json"),
-      loadContract("./artifacts/PaymentFailureTest.json"),
-    ];
-    // All artifacts are test suites.
-    const testSuites = artifacts.map((artifact) => artifact.id);
-    const config = {
-      projectRoot: __dirname,
-    };
-
-    const results = await runAllSolidityTests(
-      context,
-      OP_CHAIN_TYPE,
-      artifacts,
-      testSuites,
-      config
-    );
-
-    assert.equal(results.length, artifacts.length);
-
-    for (const res of results) {
-      if (res.id.name.includes("SetupConsistencyCheck")) {
-        assert.equal(res.testResults.length, 2);
-        assert.equal(res.testResults[0].status, "Success");
-        assert.equal(res.testResults[1].status, "Success");
-      } else if (res.id.name.includes("PaymentFailureTest")) {
-        assert.equal(res.testResults.length, 1);
-        assert.equal(res.testResults[0].status, "Failure");
-      } else {
-        assert.fail("Unexpected test suite name: " + res.id.name);
-      }
-    }
-  });
 });
-
-// Load a contract built with Hardhat into a test suite
-function loadContract(artifactPath: string): Artifact {
-  const compiledContract = require(artifactPath);
-
-  const id: ArtifactId = {
-    name: compiledContract.contractName,
-    solcVersion: "0.8.18",
-    source: compiledContract.sourceName,
-  };
-
-  const contract: ContractData = {
-    abi: JSON.stringify(compiledContract.abi),
-    bytecode: compiledContract.bytecode,
-  };
-
-  return {
-    id,
-    contract,
-  };
-}
-
-async function runAllSolidityTests(
-  context: EdrContext,
-  chainType: string,
-  artifacts: Artifact[],
-  testSuites: ArtifactId[],
-  configArgs: SolidityTestRunnerConfigArgs
-): Promise<SuiteResult[]> {
-  return new Promise((resolve, reject) => {
-    const resultsFromCallback: SuiteResult[] = [];
-
-    context
-      .runSolidityTests(
-        chainType,
-        artifacts,
-        testSuites,
-        configArgs,
-        {}, // Empty tracing config
-        (suiteResult: SuiteResult) => {
-          resultsFromCallback.push(suiteResult);
-          if (resultsFromCallback.length === artifacts.length) {
-            resolve(resultsFromCallback);
-          }
-        }
-      )
-      .catch(reject);
-  });
-}
